maas/requirements.txt
ipaddr
requests
lxml
python-cinderclient
python-glanceclient
python-heatclient
python-keystoneclient
python-neutronclient
python-novaclient
python-memcached
